backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
yt-dlp>=2024.8.6
python-multipart==0.0.9
httpx==0.27.2
openai>=1.0.0
python-dotenv>=1.0.0
